refactor(debugger): modularize TilemapViewer into sub-components
- Extract models and enums to tilemap_models.dart
- Extract CustomPainters to tilemap_painters.dart
- Extract helper widgets to tilemap_widgets.dart
- Extract Side Panel and Settings Dialog to dedicated files
- Reduce tilemap_viewer.dart size from ~2000 to ~800 lines
